[Example-Flux] Initial commit
This add all project source and resource files to a flux implementation
using kotlin for a base android application as example. This does not
perform any network operation, it, for while, just create a view action
and dispatch that action for our example store so after the store reduce
this action the store emits its change event and then ui update itself
properly.

Signed-off-by: Andre Loureiro <loureiro@work.co>